test: adiciona teste do adicionar escola planejamento
